Relocate Topbar component
