+++ src/app/[locale]/dashboard/_components/Topbar/styles.ts
